Move json mapping to Repository
