+++ app/src/main/java/com/example/cft_converter/data/repository/CurrencyRepositoryImpl.kt
@@ -3,6 +3,7 @@ package com.example.cft_converter.data.repository
 import com.example.cft_converter.data.local_data_source.LocalDataSource
 import com.example.cft_converter.data.remote_data_source.RemoteDataSource
 import com.example.cft_converter.data.mapper.CurrencyMapper
+import com.example.cft_converter.data.mapper.JsonMapper
 import com.example.cft_converter.domain.CurrencyRepository
 import com.example.cft_converter.domain.entity.CurrencyEntity
 import io.reactivex.android.schedulers.AndroidSchedulers
@@ -15,7 +16,9 @@ class CurrencyRepositoryImpl(
 ) : CurrencyRepository {
 
     override fun requestFreshListOfCurrencies(error: (Throwable) -> Unit) {
-        remoteDataSource.requestFreshListOfCurrencies({ listOfCurrencies ->
+        remoteDataSource.requestFreshListOfCurrencies({ jsonObject ->
+            val jsonMapper = JsonMapper()
+            val listOfCurrencies = jsonMapper.invoke(jsonObject)
             localDataSource.saveListOfCurrencies(listOfCurrencies)
         }, {
             error(it)
